requirements.txt
gradio
openai-whisper
pandas